requirements.txt
Django==5.2.4
django-crispy-forms
crispy-bootstrap5
psycopg2-binary
gunicorn
whitenoise